+++ src/deprecations.jl
@@ -7,3 +7,9 @@ import Base: @deprecate
 @deprecate(densematrix(args...), flatmatrix(args...))
 @deprecate(is_flat_array(args...), isflatarray(args...))
 @deprecate(has_oneto_axes(args...), has_standard_indexing(args...))
+@deprecate(SparseOperator(rowdims::Union{Integer,Tuple{Vararg{Integer}}},
+                          coldims::Union{Integer,Tuple{Vararg{Integer}}},
+                          C::AbstractVector,
+                          I::AbstractVector{<:Integer},
+                          J::AbstractVector{<:Integer}),
+           SparseOperator(C, I, J, rowdims, coldims))